Initial commit
This is the initial commit of my new new new website, andrewmiotke.com

* Using jekyll to build the site, stock minima theme
* Hosted on gh-pages
* Blog posted are nothing fancy are just markdown

Over the last few years I have changed the way I wanted to have this
website, most recently it's been a Django application. That Django
application served me well and I learned a lot of things when building
that website. It was hosted on Linode with running Ubuntu so I could to
learn a lot about building a linux web server and deploying a web app to
it.

Sadly, I've decided I didn't want to spend the $5/month on and also
continue to maintain the server. I'ts just not what I'm interested in.
I am much more interested in build applications and software on the
Apple platforms. I'm taking this time to now re-build the this website,
host it on gh-pages and not worry about it so that I can focus on
builing what I really want to be building.